feat(middlewared): invert logic to avoid an indentation level
Ticket:	#25565

+++ src/middlewared/middlewared/plugins/account.py
@@ -378,37 +378,38 @@ class UserService(CRUDService):
             )
 
     async def __update_sshpubkey(self, user, group):
-        if 'sshpubkey' in user:
-            keysfile = f'{user["home"]}/.ssh/authorized_keys'
-            pubkey = user.pop('sshpubkey')
-            if pubkey is None:
-                if os.path.exists(keysfile):
-                    try:
-                        os.unlink(keysfile)
-                    except OSError:
-                        pass
-            else:
-                oldpubkey = ''
+        if 'sshpubkey' not in user:
+            return
+        keysfile = f'{user["home"]}/.ssh/authorized_keys'
+        pubkey = user.pop('sshpubkey')
+        if pubkey is None:
+            if os.path.exists(keysfile):
                 try:
-                    with open(keysfile, 'r') as f:
-                        oldpubkey = f.read()
-                except Exception:
+                    os.unlink(keysfile)
+                except OSError:
                     pass
-                pubkey = pubkey.strip() + '\n'
-                if pubkey != oldpubkey:
-                    sshpath = f'{user["home"]}/.ssh'
-                    saved_umask = os.umask(0o77)
-                    if not os.path.isdir(sshpath):
-                        os.makedirs(sshpath)
-                    if not os.path.isdir(sshpath):
-                        raise CallError(f'{sshpath} is not a directory')
-                    if pubkey == '' and os.path.exists(keysfile):
-                        os.unlink(keysfile)
-                    else:
-                        with open(keysfile, 'w') as f:
-                            f.write(pubkey)
-                        await run('chown', '-R', f'{user["username"]}:{group}', sshpath, check=False)
-                    os.umask(saved_umask)
+        else:
+            oldpubkey = ''
+            try:
+                with open(keysfile, 'r') as f:
+                    oldpubkey = f.read()
+            except Exception:
+                pass
+            pubkey = pubkey.strip() + '\n'
+            if pubkey != oldpubkey:
+                sshpath = f'{user["home"]}/.ssh'
+                saved_umask = os.umask(0o77)
+                if not os.path.isdir(sshpath):
+                    os.makedirs(sshpath)
+                if not os.path.isdir(sshpath):
+                    raise CallError(f'{sshpath} is not a directory')
+                if pubkey == '' and os.path.exists(keysfile):
+                    os.unlink(keysfile)
+                else:
+                    with open(keysfile, 'w') as f:
+                        f.write(pubkey)
+                    await run('chown', '-R', f'{user["username"]}:{group}', sshpath, check=False)
+                os.umask(saved_umask)
 
 
 class GroupService(CRUDService):
